Summary: Add animations to the form card.
Create new file: "@/animations/formContentSlideAnimation.ts", which contains animations that slide the form in and out of the view.

Create new file: "@/animations/formScaleAnimation.ts", which contains the animations that scale the form up and down in case the content is cluttered.

Add new function to the QuestionnaireOptions component: "handleQuestionSlide", which handles the animations and the state to make for a smooth effect.

Improve the component "ScaleDownFormAnimation", by renaming it to "ScaleFormOnScrollAnimationHandler", which uses the animations from "@/animations/formScaleAnimation.ts" to de-clutter the form when needed.